refactor(cloud-ops): move to craft\cloud\ops namespace, rename preload.php to autoload.php
Migrate all cloud-ops classes from craft\cloud to craft\cloud\ops namespace
to eliminate PSR-4 collisions with the cloud plugin. Rename preload.php to
autoload.php for consistency with Composer conventions.
