+++ packages/cloud-ops/autoload.php
@@ -3,6 +3,6 @@
 if (!function_exists('craft_modify_app_config')) {
     function craft_modify_app_config(array &$config, string $appType): void
     {
-        $config = (new \craft\cloud\AppConfig($config, $appType))->getConfig();
+        $config = (new \craft\cloud\ops\AppConfig($config, $appType))->getConfig();
     }
 }
