+++ packages/cloud-ops/src/bref/handlers/CommandHandler.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace craft\cloud\bref\handlers;
+namespace craft\cloud\ops\bref\handlers;
 
 use Bref\Context\Context;
 use Bref\Event\Handler;
-use craft\cloud\bref\craft\CraftCliEntrypoint;
+use craft\cloud\ops\bref\craft\CraftCliEntrypoint;
 use InvalidArgumentException;
 
 /**
