+++ packages/cloud-ops/src/bref/handlers/JobSqsHandler.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace craft\cloud\bref\handlers;
+namespace craft\cloud\ops\bref\handlers;
 
 use Bref\Context\Context;
 use Bref\Event\Sqs\SqsEvent;
 use Bref\Event\Sqs\SqsHandler;
-use craft\cloud\bref\craft\CraftCliEntrypoint;
+use craft\cloud\ops\bref\craft\CraftCliEntrypoint;
 use RuntimeException;
 
 /**
