+++ packages/cloud-ops/src/bref/handlers/command-cmd.php
@@ -2,4 +2,4 @@
 
 require __DIR__ . '/../../../../vendor/autoload.php';
 
-return new \craft\cloud\bref\handlers\CommandHandler();
+return new \craft\cloud\ops\bref\handlers\CommandHandler();
