+++ packages/cloud-ops/src/bref/handlers/command-sqs-cmd.php
@@ -2,4 +2,4 @@
 
 require __DIR__ . '/../../../../../autoload.php';
 
-return new \craft\cloud\bref\handlers\CommandSqsHandler();
+return new \craft\cloud\ops\bref\handlers\CommandSqsHandler();
