+++ packages/cloud-ops/src/cli/controllers/BaseController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace craft\cloud\cli\controllers;
+namespace craft\cloud\ops\cli\controllers;
 
 use craft\console\Controller;
 use yii\console\Exception;
