+++ packages/cloud-ops/src/cli/controllers/InfoController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace craft\cloud\cli\controllers;
+namespace craft\cloud\ops\cli\controllers;
 
 use Craft;
 use craft\cloud\Plugin;
