+++ packages/cloud-ops/src/cli/controllers/QueueController.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace craft\cloud\cli\controllers;
+namespace craft\cloud\ops\cli\controllers;
 
 use Craft;
-use craft\cloud\queue\TestJob;
+use craft\cloud\ops\queue\TestJob;
 use craft\queue\Queue;
 use yii\console\ExitCode;
 use yii\queue\ExecEvent;
