+++ packages/cloud-ops/src/imagetransforms/ImageTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace craft\cloud\imagetransforms;
+namespace craft\cloud\ops\imagetransforms;
 
 use Craft;
 use craft\base\Component;
