+++ packages/cloud/src/Plugin.php
@@ -2,6 +2,8 @@
 
 namespace craft\cloud;
 
+use craft\cloud\ops\BasePlugin;
+
 class Plugin extends BasePlugin
 {
 }
